+++ src/api/index.ts
@@ -3,10 +3,10 @@ export * from './link'
 export * from './prune'
 
 import link from './link'
-import verify from './verify'
+import storeStatus from './storeStatus'
 import uninstall from './uninstall'
 export {
   link,
   uninstall,
-  verify,
+  storeStatus,
 }
