+++ src/cmd/store.ts
@@ -1,15 +1,18 @@
-import verify from '../api/verify'
+import storeStatus from '../api/storeStatus'
 import {PnpmOptions} from '../types'
 import {PnpmError} from '../errorTypes'
 
 export default async function (input: string[], opts: PnpmOptions) {
-  const modifiedPkgs = await verify(opts)
+  if (input[0] !== 'status') {
+    throw new Error('Unknown command')
+  }
+  const modifiedPkgs = await storeStatus(opts)
   if (!modifiedPkgs || !modifiedPkgs.length) return
 
-  throw new VerificationError(modifiedPkgs)
+  throw new StoreStatusError(modifiedPkgs)
 }
 
-class VerificationError extends PnpmError {
+class StoreStatusError extends PnpmError {
   constructor (modified: string[]) {
     super('MODIFIED_DEPENDENCY', '')
     this.modified = modified
