+++ test/index.ts
@@ -6,6 +6,6 @@ import './link'
 import './prune'
 import './cache'
 import './breakingChanges'
-import './verify'
+import './storeStatus'
 import './shrinkwrap'
 import './offline'
