+++ tsconfig.json
@@ -31,20 +31,19 @@
     "src/api/lock.ts",
     "src/api/prune.ts",
     "src/api/removeOrphanPkgs.ts",
+    "src/api/storeStatus.ts",
     "src/api/uninstall.ts",
-    "src/api/verify.ts",
     "src/bin/pnpm.ts",
     "src/binify.ts",
     "src/cmd/install.ts",
     "src/cmd/installTest.ts",
     "src/cmd/link.ts",
     "src/cmd/prune.ts",
-    "src/cmd/publish.ts",
     "src/cmd/run.ts",
     "src/cmd/runNpm.ts",
+    "src/cmd/store.ts",
     "src/cmd/uninstall.ts",
     "src/cmd/update.ts",
-    "src/cmd/verify.ts",
     "src/depsToSpecs.ts",
     "src/err.ts",
     "src/errorTypes.ts",
